Update zoom_lens.py
Added a new 'fast_init' option that uses threading to home the stages simultaneously. This takes about ~12s compared to about ~30s for the sequential home routine. For this to work the stages must be parked in fairly precise locations! To achieve this in a new setup, push the stages to ~0mm (power off) and then run with 'fast_init=False' and apply the '.close' method (which correctly parks the stages). Then start using 'fast_init=True'. Print statements were also tidied up.